requirements.txt
matplotlib
opencv-python
numpy
pytest
python-dotenv
requests
openai
google-cloud-vision
psycopg
supabase
scikit-learn
flask
gunicorn
